requirements.txt
soda-core-postgres==3.3.14
pytest==8.3.3
psycopg2-binary==2.9.9
apache-airflow-providers-postgres>=5.0.0